refactor: SummaryService
